Move the 'Create an exercise' page
- Move and rename the page component file
- Change the route path for the page
- Change associated tests
